:recycle: Functionalize cl scan label handling
